+++ examples/experimental/js/without-map/app.js
@@ -1,13 +1,11 @@
 /* global window, fetch */
-import {experimental, GeoJsonLayer} from 'deck.gl';
-const {DeckGLJS, MapControllerJS} = experimental;
+import {Deck, GeoJsonLayer} from 'deck.gl';
 
 // source: Natural Earth http://www.naturalearthdata.com/
-// via geojson.xyz
 const GEOJSON =
   'https://d2ad6b4ur7yvpq.cloudfront.net/naturalearth-3.3.0/ne_110m_admin_1_states_provinces_shp.geojson'; //eslint-disable-line
 
-const INITIAL_VIEWPORT = {
+const INITIAL_VIEW_STATE = {
   latitude: 40,
   longitude: -100,
   zoom: 3,
@@ -17,77 +15,29 @@ const INITIAL_VIEWPORT = {
 
 class App {
   constructor(props) {
-    this.state = {
-      viewport: INITIAL_VIEWPORT,
-      width: 500,
-      height: 500,
-      data: null
-    };
+    this.deck = new Deck({
+      canvas: 'deck',
+      viewState: INITIAL_VIEW_STATE
+    });
 
     fetch(GEOJSON)
       .then(resp => resp.json())
-      .then(data => this.setState({data}));
-
-    window.addEventListener('load', this.onLoad.bind(this));
-    window.addEventListener('resize', this.onResize.bind(this));
+      .then(data => this.deck.setProps({layers: this.renderLayers(data)}));
   }
 
-  setState(state) {
-    Object.assign(this.state, state);
-    this.updateLayers();
-  }
-
-  updateLayers() {
-    if (this.deckgl) {
-      this.deckgl.setProps({
-        layers: [
-          new GeoJsonLayer({
-            data: this.state.data,
-            stroked: true,
-            filled: true,
-            lineWidthMinPixels: 2,
-            getLineColor: () => [255, 255, 255],
-            getFillColor: () => [200, 200, 200]
-          })
-        ]
-      });
-    }
-  }
-
-  onResize() {
-    this.setState({
-      width: window.innerWidth,
-      height: window.innerHeight
-    });
-  }
-
-  onViewportChange(viewport) {
-    this.setState({viewport});
-    this.deckgl.setProps(viewport);
-    this.controller.setProps(viewport);
-  }
-
-  onLoad() {
-    this.onResize();
-
-    const {viewport, width, height} = this.state;
-
-    this.deckgl = new DeckGLJS({
-      ...viewport,
-      width,
-      height,
-      debug: true,
-      layers: []
-    });
-
-    this.controller = new MapControllerJS({
-      canvas: this.deckgl.canvas,
-      ...viewport,
-      width,
-      height,
-      onViewportChange: this.onViewportChange.bind(this)
-    });
+  renderLayers(data) {
+    return [
+      new GeoJsonLayer({
+        data,
+        stroked: true,
+        filled: true,
+        lineWidthMinPixels: 2,
+        opacity: 0.4,
+        getLineColor: () => [255, 100, 100],
+        getFillColor: () => [100, 100, 200]
+      })
+    ];
   }
 }
 
-new App(); // eslint-disable-line
+window.addEventListener('load', () => new App());