+++ src/react/index.js
@@ -24,4 +24,4 @@ export {default as autobind} from './utils/autobind';
 
 // TODO - React controllers should just be integrated into deck.gl API
 export {default as MapController} from './map-controller';
-export {default as OrbitController} from './experimental/orbit-controller';
+export {default as OrbitController} from './orbit-controller';